Move to numba/utils.py
